Project Blog/blog app/requirements.txt
Django==5.0.2
Pillow==9.5.0
python-dotenv==1.0.1
django-ckeditor==6.7.0 